fixed printf format strings on windows

+++ IO/TransferFunction1D.cpp
@@ -185,10 +185,10 @@ void TransferFunction1D::Set(const std::vector<unsigned char>& tf)
   assert(tfmin[1] <= tfmax[1]);
   assert(tfmin[2] <= tfmax[2]);
   assert(tfmin[3] <= tfmax[3]);
-  MESSAGE("r min/max: %zu:%zu", size_t(tfmin[0]), size_t(tfmax[0]));
-  MESSAGE("g min/max: %zu:%zu", size_t(tfmin[1]), size_t(tfmax[1]));
-  MESSAGE("b min/max: %zu:%zu", size_t(tfmin[2]), size_t(tfmax[2]));
-  MESSAGE("a min/max: %zu:%zu", size_t(tfmin[3]), size_t(tfmax[3]));
+  MESSAGE("r min/max: %hhu:%hhu", tfmin[0], tfmax[0]);
+  MESSAGE("g min/max: %hhu:%hhu", tfmin[1], tfmax[1]);
+  MESSAGE("b min/max: %hhu:%hhu", tfmin[2], tfmax[2]);
+  MESSAGE("a min/max: %hhu:%hhu", tfmin[3], tfmax[3]);
 
   for(size_t i=0; i < vColorData.size(); ++i) {
     vColorData[i] = FLOATVECTOR4(
